(2026.03.23) 에셋 및 소스 갱신
아이템 제작 시스템 구현버전

+++ 5th_6th-Team6-CH6-Project/Source/ProjectER/SkillSystem/SkillDataAsset.cpp
@@ -22,7 +22,7 @@ FGameplayAbilitySpec USkillDataAsset::MakeSpec()
     return Spec;
 }
 
-FSkillTooltipData USkillDataAsset::GetSkillTooltipData(float InLevel) const
+FSkillTooltipData USkillDataAsset::GetSkillTooltipData(int32 InLevel) const
 {
 	FSkillTooltipData Result;
